Bug 1706866 - Change radix type to uint8_t in JS::SimpleStringToBigInt(). r=jandem
There are actually several places in vm/BigIntType.h where unsigned is
used for the radix, and several other places where uint8_t is used. In the
public API, prefer the narrower type, for consistency with the API that
will be added in the next commit.

https://bugzilla.mozilla.org/show_bug.cgi?id=1706866

Differential Revision: https://phabricator.services.mozilla.com/D117191

+++ js/src/vm/BigIntType.cpp
@@ -3771,7 +3771,7 @@ BigInt* JS::StringToBigInt(JSContext* cx, Range<const char16_t> chars) {
 }
 
 static inline BigInt* SimpleStringToBigIntHelper(
-    JSContext* cx, mozilla::Span<const Latin1Char> chars, unsigned radix,
+    JSContext* cx, mozilla::Span<const Latin1Char> chars, uint8_t radix,
     bool* haveParseError) {
   if (chars.Length() > 1) {
     if (chars[0] == '+') {
@@ -3791,7 +3791,7 @@ static inline BigInt* SimpleStringToBigIntHelper(
 }
 
 BigInt* JS::SimpleStringToBigInt(JSContext* cx, mozilla::Span<const char> chars,
-                                 unsigned radix) {
+                                 uint8_t radix) {
   if (chars.empty()) {
     JS_ReportErrorNumberASCII(cx, GetErrorMessage, nullptr,
                               JSMSG_BIGINT_INVALID_SYNTAX);
